Add Miso.Style.Types.
